(#920) rename scout checks to scout status
